refactor: implement shared base class for Cardputer keyboards to unify input handling and Lvgl integration.
