+++ HalModule/Include/flx/hal/input/CardputerMatrixKeyboard.hpp
@@ -1,11 +1,6 @@
 #pragma once
 
-#include <flx/hal/DeviceBase.hpp>
-#include <flx/hal/input/IInputDevice.hpp>
-#include <lvgl.h>
-#include <mutex>
-#include <string_view>
-#include <vector>
+#include <flx/hal/input/CardputerKeyboardBase.hpp>
 
 namespace flx::hal::input {
 
@@ -13,7 +8,7 @@ namespace flx::hal::input {
  * @brief M5Stack Cardputer Matrix Keyboard Device.
  * Scans 56 keys using a 3-to-8 demux columns and 7 row input pins.
  */
-class CardputerMatrixKeyboard : public DeviceBase<IKeyboardDevice> {
+class CardputerMatrixKeyboard : public CardputerKeyboardBase {
 public:
 
 	CardputerMatrixKeyboard();
@@ -25,67 +20,13 @@ public:
 	bool start() override;
 	bool stop() override;
 
-	// ── IInputDevice ──────────────────────────────────────────────────────
-	lv_indev_t* getLvglIndev() const override;
-	int subscribeKeyEvents(KeyEventCallback cb) override;
-	void unsubscribeKeyEvents(int id) override;
+protected:
+	void updateKeyList() override;
 
 private:
-
-	struct Point2D {
-		int x;
-		int y;
-	};
-
-	struct KeyState {
-		bool tab = false;
-		bool fn = false;
-		bool shift = false;
-		bool ctrl = false;
-		bool opt = false;
-		bool alt = false;
-		bool del = false;
-		bool enter = false;
-		bool space = false;
-		std::vector<char> values;
-		std::vector<int> hidKey;
-
-		void reset() {
-			tab = false;
-			fn = false;
-			shift = false;
-			ctrl = false;
-			opt = false;
-			alt = false;
-			del = false;
-			enter = false;
-			space = false;
-			values.clear();
-			hidKey.clear();
-		}
-	};
-
 	void initHardware();
 	void setDemuxOutput(uint8_t output);
 	uint8_t readRowInputs();
-	void updateKeyList();
-	void updateKeysState();
-
-	lv_indev_t* m_indev = nullptr;
-	std::mutex m_mutex;
-	std::vector<std::pair<int, KeyEventCallback>> m_observers;
-	int m_nextObserverId = 1;
-
-	std::vector<Point2D> m_keyListBuffer;
-	std::vector<Point2D> m_keyValuesWithoutSpecialKeys;
-	KeyState m_keysStateBuffer;
-	uint8_t m_lastKeySize = 0;
-	uint32_t m_lastKeyPressed = 0;
-	bool m_isPressed = false;
-
-	static void indevReadCb(lv_indev_t* indev, lv_indev_data_t* data);
-	void readInput(lv_indev_data_t* data);
-	void notifyObservers(const KeyEvent& event);
 };
 
 } // namespace flx::hal::input
